[cli] put the config inside context, remove options layer inside config
