+++ apps/src/lib/cli/utils.rs
@@ -7,7 +7,6 @@ use clap::ArgMatches;
 
 use super::args;
 use super::context::{Context, FromContext};
-use crate::wallet::Wallet;
 
 // We only use static strings
 pub type App = clap::App<'static>;
@@ -23,11 +22,7 @@ pub trait Cmd: Sized {
         match Self::parse(&matches) {
             Some(cmd) => {
                 let global_args = args::Global::parse(&matches);
-                let wallet = Wallet::load_or_new(&global_args.base_dir);
-                let context = Context {
-                    global_args,
-                    wallet,
-                };
+                let context = Context::new(global_args);
                 (cmd, context)
             }
             None => {
